// src/controllers/rest-action.enum.ts
export enum RestAction {
    Get = 'get',
    GetOne = 'getOne',
    Post = 'post',
    Put = 'put',
    Patch = 'patch',
    Delete = 'delete',
    BatchGet = 'batchGet',
    BatchPost = 'batchPost',
    BatchPut = 'batchPut',
    BatchPatch = 'batchPatch',
    BatchDelete = 'batchDelete',
}